requirements.txt
Flask>=2.0
Flask-JWT-Extended>=4.3

fastapi~=0.115.12
pydantic~=2.10.6
SQLAlchemy~=2.0.39
PyJWT~=2.10.1
jose~=1.0.0
passlib~=1.7.4
Flask-Login~=0.6.3